grading/llm_quiz/requirements.txt
dspy-ai>=2.4.0
tomli>=2.0.1; python_version < "3.11"